Normalizing on LESS with a compiler. Checking in newly-compiled css files.
I feel like this needs to be revisited later. With a compiler, css files
are reduced to build artifacts that should probably not be checked into
source. But for this to happen, I need a way to show others how to build
the files correctly -- IE a build script. I may bring this in when I am
closer to launch.